Add tests for stringify/normalize + bug fix
