+++ packages/preserve/.gitignore
@@ -2,3 +2,4 @@ node_modules/*
 yarn.lock
 package-lock.json
 dist
+coverage
