+++ packages/preserve/lib/targets/sources/types.ts
@@ -12,7 +12,10 @@ export interface Container {
 export type Source = Content | Container;
 
 export const isContainer = (source: Source): source is Container =>
-  source && typeof source === "object" && "entries" in source;
+  source &&
+  typeof source === "object" &&
+  "entries" in source &&
+  typeof source.entries === "object";
 
 export const isContent = (source: Source): source is Content =>
   !isContainer(source);
